Add support for uploading files from local device
