+++ ui/.env.example
@@ -1,10 +1,16 @@
 NEXTAUTH_URL="http://localhost:3000"
 NEXTAUTH_SECRET="superagent"
 NEXT_PUBLIC_SUPERAGENT_API_URL="http://127.0.0.1:8080/api/v1"
+# Mandatory for uploading files
+NEXT_PUBLIC_AWS_S3_BUCKET=s
+NEXT_PUBLIC_AWS_S3_REGION=
+NEXT_PUBLIC_AMAZON_S3_ACCESS_KEY_ID=
+NEXT_PUBLIC_AMAZON_S3_SECRET_ACCESS_KEY=
+
 # Shareable Key Needs to be exactly 16 characters, do not wrap value in quotes
 NEXT_PUBLIC_SHARABLE_KEY_SECRET=
 
-# Optiona OAuth values
+# Optional OAuth values
 NEXT_PUBLIC_GITHUB_CLIENT_ID=""
 NEXT_PUBLIC_GITHUB_CLIENT_SECRET=""
 NEXT_PUBLIC_GOOGLE_CLIENT_ID=""
