+++ ui/app/datasources/_components/document-picker.js
